Cleaned up a lot of code by separating out axios requests into meetupService.js
