+++ controllers/auth_controller.js
@@ -7,7 +7,7 @@ const redirect_uri = process.env.REDIRECT_URI;
 const User = require("./../database/models/user_model");
 
 //Packages / Imports
-require("dotenv").config();
+
 const axios = require('axios');
 const queryString = require('query-string');
 const meetupService = require('../services/meetupService');
@@ -35,103 +35,39 @@ async function meetupAuth (req, res) {
         
         console.log('This is the meetup authorization js file.')
 
-        //Define the headers as using URL-encoded format
-        const config = { headers: {
-            'Content-Type': 'application/x-www-form-urlencoded'
-        }};
-    
-        //Client auth information to be used in the body.
-        //Use queryString.stringify to convert this information to url-encoded from JSON.
-        const body = queryString.stringify({
-            'client_id': process.env.CLIENT_ID,
-            'client_secret':process.env.CLIENT_SECRET,
-            'grant_type': 'authorization_code',
-            'redirect_uri': process.env.REDIRECT_URI,
-            'code':req.query.code
-        });
-    
-    
-       try {
-        //If auth successful, provide the response (including access and refresh token)
-        const response = await axios.post
-        (
-            'https://secure.meetup.com/oauth2/access',
-            body,
-            config
-        );
-
-        // //Use a singleton pattern to store the tokens
-        // //THIS IS A TEMPORARY SOLUTION - will store this info in the database later
-        // //TODO - replace 'current-user' with user ID from database
-        // meetupService.setItem("current-user", {
-        //     "access_token": response.data.access_token,
-        //     "refresh_token": response.data.refresh_token
-        // });
-
+       const tokens = await meetupService.getTokens(req.query.code);
+       const userData = await meetupService.getUserInfo(tokens.access_token);
         //axios request for user data.
         //store it in a var
         // check if it exists in the db
         // if not, create
         // if it does, update
 
-        axios.get('https://api.meetup.com/members/self', {
-            headers: {
-                "Authorization": `Bearer ${response.data.access_token}`
-            }
-        }).then (function(response) {
+        console.log(userData);
 
             //Take the params that meetup allows us to take and store this in an object
             let userProfileInfo = {
-                meetup_uid: response.data.id,
-                email: response.data.email,
-                firstName: response.data.name,
-                city: response.data.city,
-                avatar: response.data.photo.photo_link,
-                access_token: response.data.access_token,
-                refresh_token: response.data.refresh_token
+                meetup_uid: userData.id,
+                email: userData.email,
+                firstName: userData.name,
+                city: userData.city,
+                avatar: userData.photo.photo_link,
+                access_token: userData.access_token,
+                refresh_token: userData.refresh_token
             }
 
-            const user = 
-            User.find({userProfileInfo}
-            ).then(
-                function(user) {
-                    console.log('found')
-                }
-            ).catch(
-                function(err) {
-                    console.log(err);
-                }
-            )
-            
+            //Does the user exist?
+            const user =  await User.findOne({"meetup_uid": userProfileInfo.meetup_uid})
+            if (user) {
+                //update
+            } else {
+                //create
+            }
+        
 
-            // // //Does the user exist in the database?
-            // if (user) {
-            //     console.log('user found');
-            // } else {
-            //     console.log ('not found');
-            // }
-
-
-
-
-        }).catch (function(error) {
-            console.log(error);
-        })
-
-        const userDetails = {
-
-        }
-
-        console.log("Access token: " + response.data.access_token);
         return res.redirect("/");
     } 
-    
-       catch(err) {
-        console.log(err);
-       }
-
-
-    }
+
